idml: the writer's "4 decimals" is a contract at every magnitude again
`format_f32` rounded by computing `(v * 10_000.0).round()` in `f32`. That
multiply is never exact — an `f32` significand is 24 bits and the product
needs up to 34 — so the value `.round()` sees has already slipped. Below
`2^24 / 10_000 = 1677.7216` the slip can only carry the product across one
`.5` boundary, so the damage is bounded at a single wrong ten-thousandth.
At and above that magnitude the product passes `2^24`, where consecutive
representable `f32` are more than 1 apart: `.round()` stops rounding
anything at all, and the documented "4 decimals" quietly becomes coarser
with no announcement.

MEASURED, NOT DERIVED ON PAPER. Walking every `f32` rather than sampling:
below the wall nothing is ever printed more than one ten-thousandth away
(the nearest value that is off at all is `1677.6563`, by one). The first
`f32` above the wall, `1677.7217`, already collides with its own
neighbour — both print `1677.7216` — and `1677.7225` is the first printed
more than one ten-thousandth away. Worst error per band, in units of the 1e-4
being claimed: 2 at [1677.7216, 2048), 3 at [2048, 4096), 10 at
[8192, 16384), 79 at [65536, 131072). 14 % of all values above the wall
are affected.

WHY IT IS NOT ONLY A BYTE-IDENTITY CONCERN. This function emits values on
MUTATED saves — a dragged item's `ItemTransform`, a resized frame's
`GeometricBounds` — and an InDesign pasteboard runs to five figures of
points, so it is reachable by editing. It moves the corpus save-back
number not at all (57 before, 57 after: no unmutated corpus value crosses
the wall on a re-derived attribute), which is precisely why it needed
justifying on its own terms rather than by the gap count.

WHAT IT DOES AND DOES NOT RECOVER, because overclaiming here would be
worse than the bug. Widening the ROUNDING STEP to `f64` makes the multiply
exact (34 bits fits in 53), so the digits printed are the correctly-rounded
ones at every magnitude — verified against an integer-arithmetic reference
over every `f32` in [0, 1), [1024, 4096), [16384, 32768) and [1e6, 2e6),
1.09 billion values, zero mismatches. Above the wall that is exactly
lossless: 1e-4 is FINER than the input's own ULP there, so the emitted
spelling re-parses to the very same `f32`, where the old one lost up to a
full ULP.

It buys no precision the `f32` never had. The input is `f32` and stays
`f32`; above the wall its own spacing is COARSER than the ten-thousandth
being printed (ten quanta at 13 kpt), so most 4-decimal spellings are
unreachable and the number was already that coarse before it reached the
formatter. Below the wall, 4 decimals still discards most of what an `f32`
holds, exactly as before — that is InDesign's output format, and the
answer to it is preserving a source spelling rather than re-deriving one,
which is a different lane and only untouched values get it.

Nor is the practical magnitude dramatic: the worst movement a save
introduced was 2.4e-4 pt at 3 kpt and 7.8e-3 pt at 107 kpt, i.e. sub-micron
to ~3 µm, and it is one-shot rather than accumulating — repeated
open/save reaches a fixed point after one cycle at every magnitude tested.
The defect is that a documented contract was false and degraded silently,
not that pages came out visibly wrong.

`paged_gen::xml::format_f32` in core, which this is a copy of, still has
the f32 rounding; out of scope here, worth a look there.

+6 tests (249 -> 255), three of which fail at the parent commit. They run
through the public `rewrite_spread` — a drag and a resize to pasteboard
coordinates — rather than reaching for the private formatter, and two of
them exist to pin the LIMITS of the claim above so a later reader does not
mistake this for a precision upgrade.

+++ crates/idml-export/src/rewrite.rs
@@ -222,8 +222,32 @@ use idml_import::{Bounds, CharacterRun, PathAnchor, Spread, Story, TableCell, Te
 /// crate that pulls clap/anyhow) so this runtime crate stays minimal +
 /// wasm-clean. InDesign serialises floats this way, so patched values
 /// match the surrounding hand-written / exported numbers.
+///
+/// # Why the rounding step is `f64`
+///
+/// `paged-gen`'s original rounds in `f32` — `(v * 10_000.0).round()` —
+/// and that multiply is never exact: an `f32` significand is 24 bits and
+/// the product needs up to 34, so the value `.round()` sees has already
+/// slipped. Below `2^24 / 10_000 = 1677.7216` the slip only ever moves
+/// the result across one `.5` boundary, i.e. a wrong LAST decimal
+/// (0.06 % of values around 1 pt, rising to 17 % around 300 pt). At
+/// and above that magnitude the product passes `2^24`, its own spacing
+/// exceeds 1, and `.round()` stops rounding anything at all: the
+/// documented "4 decimals" silently becomes coarser, by up to 3
+/// ten-thousandths at 2–4 kpt and 79 at 65–131 kpt. Real documents run
+/// there — an InDesign pasteboard coordinate is routinely five figures —
+/// and this function emits MUTATED values, so it is not only a
+/// byte-identity concern.
+///
+/// Widening to `f64` first makes the multiply exact (34 bits fits
+/// comfortably in 53), so the digits printed are the correctly-rounded
+/// ones at every magnitude. It buys precision the OUTPUT was throwing
+/// away, not precision the `f32` never had: above 1677.7216 pt the
+/// input's own ULP is already coarser than the 1e-4 being printed, and
+/// below it 4 decimals still discards most of what an `f32` holds. See
+/// `tests/format_f32_precision.rs`.
 pub(crate) fn format_f32(v: f32) -> String {
-    let rounded = (v * 10_000.0).round() / 10_000.0;
+    let rounded = (f64::from(v) * 10_000.0).round() / 10_000.0;
     if rounded == 0.0 {
         return "0".to_string();
     }
